refactor(composition): group automation cluster under automation/ (dissection n9)
