+++ crates/ironclaw_reborn_composition/src/automation/facade/tests.rs
