+++ crates/ironclaw_reborn_composition/src/automation/trigger_poller.rs
@@ -24,10 +24,10 @@ use tokio_util::sync::CancellationToken;
 use crate::runtime_input::TriggerPollerSettings;
 #[cfg(feature = "slack-v2-host-beta")]
 use crate::slack::slack_delivery::PostSubmitDeliveryHook;
-pub(crate) use crate::trigger_poller_trusted_submit::AccessCheckerTriggerFireAuthorizer;
-pub(crate) use crate::trigger_poller_trusted_submit::ConversationContentRefMaterializer;
+pub(crate) use crate::automation::trigger_poller_trusted_submit::AccessCheckerTriggerFireAuthorizer;
+pub(crate) use crate::automation::trigger_poller_trusted_submit::ConversationContentRefMaterializer;
 #[cfg(any(test, feature = "test-support"))]
-pub(crate) use crate::trigger_poller_trusted_submit::TenantScopedTrustedTriggerFireAuthorizer;
+pub(crate) use crate::automation::trigger_poller_trusted_submit::TenantScopedTrustedTriggerFireAuthorizer;
 
 mod active_run_lookup;
 pub(crate) use active_run_lookup::SnapshotActiveRunLookup;
