+++ crates/ironclaw_reborn_composition/src/automation/trigger_poller_trusted_submit.rs
