+++ crates/ironclaw_reborn_composition/src/test_support/automation.rs
@@ -15,7 +15,7 @@ use ironclaw_triggers::TriggerRepository;
 pub fn local_dev_automation_product_facade_for_test(
     trigger_repository: Arc<dyn TriggerRepository>,
 ) -> Arc<dyn AutomationProductFacade> {
-    Arc::new(crate::automation::RebornAutomationProductFacade::new(
+    Arc::new(crate::automation::facade::RebornAutomationProductFacade::new(
         trigger_repository,
     ))
 }
